move specialized equations to zoo
