+++ docs/source/_rst/equation/zoo/acoustic_wave_equation.rst
@@ -1,0 +1,7 @@
+AcousticWaveEquation
+=====================
+.. currentmodule:: pina.equation.zoo.acoustic_wave_equation
+
+.. automodule:: pina._src.equation.zoo.acoustic_wave_equation
+    :members:
+    :show-inheritance:
