+++ docs/source/_rst/equation/zoo/advection_equation.rst
@@ -1,0 +1,7 @@
+Advection Equation
+=====================
+.. currentmodule:: pina.equation.zoo.advection_equation
+
+.. automodule:: pina._src.equation.zoo.advection_equation
+    :members:
+    :show-inheritance:
