+++ docs/source/_rst/equation/zoo/allen_cahn_equation.rst
@@ -1,0 +1,7 @@
+Allen Cahn Equation
+=====================
+.. currentmodule:: pina.equation.zoo.allen_cahn_equation
+
+.. automodule:: pina._src.equation.zoo.allen_cahn_equation
+    :members:
+    :show-inheritance:
