+++ docs/source/_rst/equation/zoo/diffusion_reaction_equation.rst
@@ -1,0 +1,7 @@
+Diffusion Reaction Equation
+==============================
+.. currentmodule:: pina.equation.zoo.diffusion_reaction_equation
+
+.. automodule:: pina._src.equation.zoo.diffusion_reaction_equation
+    :members:
+    :show-inheritance:
