+++ docs/source/_rst/equation/zoo/helmholtz_equation.rst
@@ -1,0 +1,9 @@
+Helmholtz Equation
+=====================
+.. currentmodule:: pina.equation.zoo.helmholtz_equation
+
+.. automodule:: pina._src.equation.zoo.helmholtz_equation
+
+.. autoclass:: pina._src.equation.zoo.helmholtz_equation.HelmholtzEquation
+    :members:
+    :show-inheritance:
