+++ docs/source/_rst/equation/zoo/poisson_equation.rst
@@ -1,0 +1,9 @@
+Poisson Equation
+=====================
+.. currentmodule:: pina.equation.zoo.poisson_equation
+
+.. automodule:: pina._src.equation.zoo.poisson_equation
+
+.. autoclass:: pina._src.equation.zoo.poisson_equation.PoissonEquation
+    :members:
+    :show-inheritance:
